+++ packages/react18-themes/src/server/nextjs/nextjs-ssg-theme-switcher/nextjs-ssg-theme-switcher.tsx
@@ -8,17 +8,16 @@ export type ForcedPage = [pathMatcher: RegExp | string, themes: { theme?: string
 
 export interface NextJsSSRThemeSwitcherProps extends HTMLProps<HTMLElement> {
 	children?: ReactNode;
-	/** @default 'div' */
+	/** @defaultValue 'div' */
 	tag?: keyof JSX.IntrinsicElements;
 	forcedPages?: ForcedPage[];
 }
 
-/**
- * Server side wrapper for Next.js to replace &#x60;html&#x60; tag
- */
-export function NextJsSSRThemeSwitcher({ children, tag, forcedPages, ...props }: NextJsSSRThemeSwitcherProps) {
-	const Tag: keyof JSX.IntrinsicElements = tag || "div";
-
+function sharedServerComponentRenderer(
+	{ children, tag, forcedPages, ...props }: NextJsSSRThemeSwitcherProps,
+	defaultTag: "div" | "html",
+) {
+	const Tag: keyof JSX.IntrinsicElements = tag || defaultTag;
 	const state = cookies().get("react18-themes")?.value;
 
 	const path = headers().get("referer");
@@ -37,10 +36,25 @@ export function NextJsSSRThemeSwitcher({ children, tag, forcedPages, ...props }:
 
 	return (
 		// @ts-expect-error -> svg props and html element props conflict
-		<Tag id="react18-themes" {...dataProps} {...props} data-testid="nextjs-ssr-theme-switcher">
+		<Tag id="react18-themes" {...dataProps} {...props} data-testid="nextjs-server-theme-switcher">
 			{children}
 		</Tag>
 	);
+}
+
+export function NextJsSSGThemeSwitcher(props: NextJsSSRThemeSwitcherProps) {
+	return sharedServerComponentRenderer(props, "div");
+}
+
+export interface ServerSideWrapperProps extends NextJsSSRThemeSwitcherProps {
+	/** @defaultValue 'html' */
+	tag?: keyof JSX.IntrinsicElements;
+}
+/**
+ * Server side wrapper for Next.js to replace &#x60;html&#x60; tag
+ */
+export function ServerSideWrapper(props: ServerSideWrapperProps) {
+	return sharedServerComponentRenderer(props, "html");
 }
 
 interface Theme {